celeb intro: runs if we don't know fav celeb, we haven't just asked for their fav celeb, and we weren't told to move onto a non-celeb topic
celeb unknown: runs in two cases. first, we just asked for fav celeb, and we dont know who they were talking abt. second, we don't know one of the celeberity's works that the user mentioned. also, in both cases, we can't have been told to move onto a non-celeb topic.

celeb discuss: runs with more priority than celeb generic. runs only if we havent prompted for current topic, we have just asked for user's fav celeb, we are able to discuss this celeb's works, and we were not told to move otno a non-celeb topic.

celeb generic: only runs if we havent prompted for current topic, we have an opinion in our DB, and we were not told to move onto a new topic.